Move aggregated bytes to io
