+++ aws-s3-transfer-manager/src/operation/download/service.rs
@@ -4,6 +4,7 @@
  */
 use crate::error;
 use crate::http::header;
+use crate::io::aggregated_bytes::AggregatedBytes;
 use crate::middleware::limit::concurrency::ConcurrencyLimitLayer;
 use crate::middleware::retry;
 use crate::operation::download::DownloadContext;
@@ -17,7 +18,6 @@ use tokio::task;
 use tower::{service_fn, Service, ServiceBuilder, ServiceExt};
 use tracing::Instrument;
 
-use super::body::AggregatedBytes;
 use super::body::ChunkResponse;
 use super::{DownloadInput, DownloadInputBuilder};
